requirements.txt
numpy
pandas
matplotlib
plotly
scikit-learn
tensorflow>=2.13
torch>=2.1
torchvision>=0.16
torchaudio>=2.1
shap>=0.46.0
ta
streamlit>=1.36
pyyaml
joblib
